requirements.txt
fastapi[standard]
langchain
ipykernel
python-dotenv
langchain_community
pypdf
pymupdf
bs4
wikipedia
langchain_text_splitters
langchain-openai
faiss-cpu
chromadb
langchain_ollama
langchain_groq
fastapi
uvicorn
tiktoken
PyPDF2
langchain_nvidia_ai_endpoints
pinecone
pinecone_text
langchain_nvidia
langchain_pinecone
pydantic
langchain_core
streamlit
